requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
# transformers==4.40.0
# torch>=2.0.0
pydantic==2.6.4
sqlalchemy>=2.0.0
pyodbc>=5.0.0
gunicorn
python-dotenv>=1.0.0
httpx>=0.27.0
